Move execute()/start() methods up
